feat: Monorepo setup
Due to strict and customized expo dependencies, most classic packages conflict with expo.

Expo does not seem to be suitable for a monorepo environment. It is therefore to be expected that there will be common conflicts when installing or updating packages. It is preferable to abandon expo to use another technology (react-native vanilla, ionic, cordova + react, phonegap + react).